refactor(auth): Extracted fetch user to the user service
